backend/requirements.txt
fastapi
uvicorn[standard]
crawl4ai
anthropic
sqlalchemy[asyncio]
asyncpg
python-dotenv
httpx
pydantic
